refactor: centralize types into src/types/ (ipc, file, viewer, ui)
