+++ docs/REFACTORING-PLAN.md
@@ -237,7 +237,7 @@ Phase 3-8 は Phase 1-2 と並行して進めても問題ない。
 | 1.2 lib.rs 分割      | 🟢   | 2026-05-29 | lib.rs 2,458→178行 (-92.8%)。commands/ (files/settings/diagnostics/usd/updater/alembic/integrations/bench/shot), state.rs, shared.rs に分割。cargo check + test 通過。                                                                                                                                                                                                                              |
 | 2 状態管理整理       | 🟢   | 2026-05-29 | zustand 導入。App.tsx の62個の useState を useViewerStore / useFileStore / useUiStore に移行。toggle/update アクション追加。App.tsx 1,919→1,838行 (このフェーズでは行数削減より構造改善が主目的)。                                                                                                                                                                                                  |
 | 3 CSS整理            | 🟢   | 2026-05-29 | design-system.css に `:root` 一本化 + `--font-mono`/`--font-sans` エイリアス追加。styles.css の hex 値をトークンに置換。`.btn-ghost`/`.btn-primary` を `<Button>` に移行→削除。未使用 CSS 削除（`.card-header-count`, `.extension-badges`, `.error-text`, `.warning-text`, `.viewer-state-warning`, `.viewer-state-danger`, `.range-control`, `.loader` 一式）。styles.css 2,436→2,124行 (-12.8%)。 |
-| 4 型の一元管理       | ⏳   | —          |                                                                                                                                                                                                                                                                                                                                                                                                     |
+| 4 型の一元管理       | 🟢   | 2026-05-29 | `src/types/` に ipc.ts (266行), file.ts (42行), viewer.ts (487行), ui.ts (122行) を作成。IPCワイヤ型、ファイル型、ビューワ/メタデータ型、UI状態型を一元管理。元ファイルに re-export を追加し、既存インポートの互換性を維持。tsc + test 197件 通過。 |
 | 5 エラーハンドリング | ⏳   | —          |                                                                                                                                                                                                                                                                                                                                                                                                     |
 | 6 テスト拡充         | ⏳   | —          |                                                                                                                                                                                                                                                                                                                                                                                                     |
 | 7 ビルド/CI最適化    | ⏳   | —          |                                                                                                                                                                                                                                                                                                                                                                                                     |
