+++ src/components/assetMetadata.ts
@@ -1,247 +1,20 @@
-export type HierarchyNode = {
-  name: string;
-  displayName?: string;
-  kind: string;
-  children: HierarchyNode[];
-  /** #46: full USD SdfPath surfaced from GLB node extras.primPath.
-   * Present only when the asset went through the hierarchy-aware GLB
-   * pipeline. `undefined` for non-USD assets (GLTF/FBX/OBJ). */
-  primPath?: string;
-};
+import type { AssetMetadata } from "../types/viewer";
 
-export type TextureEntry = {
-  id: string;
-  label: string;
-  channel: string;
-  dimensions: string;
-  thumbnailUrl: string | null;
-  previewFlipY?: boolean;
-  sourceKind: "embedded" | "external" | "standalone" | "unresolved" | "unknown";
-};
-
-/** A texture slot in a material, carrying only the name so the inspector
- * can display the source without needing to transfer the GPU texture object
- * across the viewer boundary. */
-export type MaterialTextureSlot = {
-  /** Three.js `Texture.name`, or the userData-derived path when available.
-   * Falls back to the slot label (e.g. `"Base Color"`) when no name is
-   * present on the texture. */
-  name: string;
-};
-
-export type MmdMaterialEntry = {
-  materialIndex: number | null;
-  name: string;
-  englishName: string | null;
-  diffuse: [number, number, number, number] | null;
-  specular: [number, number, number] | null;
-  ambient: [number, number, number] | null;
-  specularPower: number | null;
-  edgeColor: [number, number, number, number] | null;
-  edgeSize: number | null;
-  texturePath: string | null;
-  sphereTexturePath: string | null;
-  sphereMode: string | null;
-  toonTexturePath: string | null;
-  sharedToonIndex: number | null;
-  transparencyMode: string | null;
-  renderOrderBucket: string | null;
-  faceCount: number | null;
-  flags: Record<string, boolean> | null;
-  unsupportedDrawFlags: string[];
-};
-
-export type MaterialEntry = {
-  id: string;
-  name: string;
-  type: string;
-  color: string | null;
-  opacity: number;
-  transparent: boolean;
-  textureCount: number;
-  /** Names of meshes that bind this material. Derived from the live
-   * Three.js scene graph because the GLB → glTF → Three.js round-trip
-   * drops authored USD prim paths; mesh names are the closest stand-in
-   * the inspector can recover. Multiple meshes may share a material,
-   * and a single mesh authoring an array material appears once per
-   * slot. Empty when no mesh references the material (rare). */
-  boundMeshes: string[];
-
-  // ── Shader input slot detail (Issue #36) ──────────────────────────────
-  /** RGBA base color factor. Alpha is derived from `Material.opacity`.
-   * Present only for `MeshStandardMaterial` / `MeshPhongMaterial` /
-   * `MeshBasicMaterial`; `null` otherwise. */
-  baseColorFactor: [number, number, number, number] | null;
-  /** `MeshStandardMaterial.metalness`. `null` for non-standard materials. */
-  metallicFactor: number | null;
-  /** `MeshStandardMaterial.roughness`. `null` for non-standard materials. */
-  roughnessFactor: number | null;
-  /** RGB emissive factor. Present for Standard / Phong; `null` otherwise. */
-  emissiveFactor: [number, number, number] | null;
-  /** Base-color / albedo texture, or `null` when none is assigned. */
-  baseColorTexture: MaterialTextureSlot | null;
-  /** Combined metallic-roughness texture (`metalnessMap`), or `null`. */
-  metallicRoughnessTexture: MaterialTextureSlot | null;
-  /** Normal map texture, or `null`. */
-  normalTexture: MaterialTextureSlot | null;
-  /** Emissive texture, or `null`. */
-  emissiveTexture: MaterialTextureSlot | null;
-  /** Alpha handling mode inferred from Three.js material flags and
-   * `userData.gltfAlphaMode` when the glTF extras carry it. */
-  alphaMode: "OPAQUE" | "MASK" | "BLEND" | "unknown";
-  /** USD prim path surfaced through `material.userData.usdPrimPath` when
-   * the asset went through the Phase-7 USD→GLB pipeline. `null` when the
-   * round-trip drops the prim path (the common case for pure-GLB assets). */
-  usdPrimPath: string | null;
-  /** MMD material metadata from `@yohawing/three-mmd-loader`, when present. */
-  mmd: MmdMaterialEntry | null;
-};
-
-/** One light surfaced in the scene panel. Authored by USD as
- * `UsdLuxDistantLight` / `SphereLight` / etc. and lowered to a
- * Three.js `Light` instance by GLTFLoader's `KHR_lights_punctual`
- * extension. The frontend re-derives this from the live scene graph
- * because Three.js loses the original USD prim path during the
- * GLB → glTF → Three.js round-trip. */
-export type LightEntry = {
-  id: string;
-  name: string;
-  /** Three.js `light.type` (e.g. `"DirectionalLight"`,
-   * `"PointLight"`, `"SpotLight"`, `"AmbientLight"`,
-   * `"HemisphereLight"`). */
-  type: string;
-  /** Hex color of the emitted light, or `null` when the light type
-   * does not expose a color (rare). */
-  color: string | null;
-  /** glTF `KHR_lights_punctual` intensity in candela (point/spot) or
-   * lux (directional). Three.js stores this directly on `light.intensity`. */
-  intensity: number;
-};
-
-/** One camera authored on the stage. USD `UsdGeomCamera` prims are
- * lowered to glTF `cameras` by the Phase 7b backend pass, and
- * GLTFLoader instantiates them as Three.js `PerspectiveCamera` (or
- * `OrthographicCamera` once Phase 7b ortho support lands) attached to
- * the scene graph. Switching the active camera is a separate task —
- * this entry exists so the inspector can list authored cameras even
- * while the viewer keeps using its own orbit camera. */
-export type CameraEntry = {
-  id: string;
-  name: string;
-  /** `"perspective"` for `PerspectiveCamera`, `"orthographic"`
-   * otherwise. */
-  projection: "perspective" | "orthographic";
-  /** Vertical field of view in degrees. `null` for orthographic. */
-  fov: number | null;
-  /** Width / height ratio reported by Three.js. `null` for orthographic. */
-  aspect: number | null;
-  near: number;
-  far: number;
-};
-
-export type MmdSectionEntry = {
-  name: string;
-  count: number;
-  offset: number;
-  byteLength: number;
-};
-
-export type MmdAssetMetadata = {
-  format: "pmx" | "pmd";
-  version: number;
-  encoding: string | null;
-  name: string;
-  englishName: string;
-  comment: string;
-  englishComment: string;
-  counts: Record<string, number>;
-  additionalUvCount: number | null;
-  indexSizes: Record<string, number> | null;
-  trailingBytes: number;
-  sections: MmdSectionEntry[];
-  diagnostics: Array<{
-    level: "warning" | "error";
-    code: string;
-    message: string;
-  }>;
-};
-
-export type MmdBoneEntry = {
-  boneIndex: number | null;
-  parentIndex: number | null;
-  parentName: string | null;
-  name: string | null;
-  englishName: string | null;
-  restPosition: [number, number, number] | null;
-  layer: number | null;
-  appendTransform: {
-    parentIndex: number;
-    parentName: string | null;
-    weight: number;
-  } | null;
-  flags: Record<string, boolean> | null;
-  ik: {
-    roles: string[];
-    goalBoneIndex: number | null;
-    effectorBoneIndex: number | null;
-    iterationCount: number | null;
-    maxAnglePerIteration: number | null;
-    linkCount: number | null;
-    limitKinds: string[];
-  } | null;
-};
-
-export type MmdMorphEntry = {
-  name: string | null;
-  englishName: string | null;
-  type: string | null;
-  boneOffsetCount: number;
-  groupOffsetCount: number;
-  flipOffsetCount: number;
-  impulseOffsetCount: number;
-};
-
-export type ObjectInfo = {
-  name: string;
-  kind: string;
-  visible: boolean;
-  position: [number, number, number];
-  rotation: [number, number, number];
-  scale: [number, number, number];
-  boundingBox: [number, number, number, number, number, number] | null;
-  vertexCount: number | null;
-  triangleCount: number | null;
-  materialNames: string[];
-  materialIds: string[];
-  morphTargets: MorphTargetEntry[];
-  childCount: number | null;
-  animatesWithClips: string[];
-  userData: Record<string, unknown> | null;
-  mmdBone: MmdBoneEntry | null;
-};
-
-export type MorphTargetEntry = {
-  index: number;
-  name: string;
-  value: number;
-  mmd: MmdMorphEntry | null;
-};
-
-export type AssetMetadata = {
-  formatLabel: string;
-  formatVersion: string | null;
-  nodeCount: number;
-  meshCount: number;
-  materialCount: number;
-  textureCount: number;
-  hasAnimation: boolean;
-  hierarchy: HierarchyNode[];
-  textures: TextureEntry[];
-  materials: MaterialEntry[];
-  lights: LightEntry[];
-  cameras: CameraEntry[];
-  objectInfo: Record<string, ObjectInfo>;
-  mmd?: MmdAssetMetadata;
-};
+export type {
+  HierarchyNode,
+  TextureEntry,
+  MaterialTextureSlot,
+  MmdMaterialEntry,
+  MaterialEntry,
+  LightEntry,
+  CameraEntry,
+  MmdSectionEntry,
+  MmdAssetMetadata,
+  MmdBoneEntry,
+  MmdMorphEntry,
+  ObjectInfo,
+  MorphTargetEntry,
+  AssetMetadata,
+} from "../types/viewer";
 
 export const emptyAssetMetadata: AssetMetadata | null = null;
