+++ src/lib/recentFiles.ts
@@ -1,15 +1,8 @@
 import { invoke } from "@tauri-apps/api/core";
 
-export type RecentFileEntry = {
-  path: string;
-  kind: string;
-  lastAccessedAt: string;
-};
+import type { RecentFilesPayload } from "../types/file";
 
-export type RecentFilesPayload = {
-  recentFilesPath: string;
-  entries: RecentFileEntry[];
-};
+export type { RecentFileEntry, RecentFilesPayload } from "../types/file";
 
 export async function loadRecentFiles() {
   return invoke<RecentFilesPayload>("load_recent_files");
