+++ src/types/viewer.ts
@@ -1,0 +1,570 @@
+import type {
+  AnimationAction,
+  AnimationClip,
+  AnimationMixer,
+  Group,
+  Mesh,
+  Object3D,
+  PerspectiveCamera,
+  PMREMGenerator,
+  Scene,
+  Texture,
+  WebGLRenderer,
+  ColorSpace,
+  ToneMapping,
+} from "three";
+import type { OrbitControls } from "three/examples/jsm/controls/OrbitControls.js";
+
+// ── Viewer mode & feedback ───────────────────────────────────────
+
+export type ViewerMode =
+  | "empty"
+  | "loading"
+  | "ready"
+  | "unsupported"
+  | "missingOptionalLoader"
+  | "loadFailed"
+  | "missingReference";
+
+export type ViewerFeedback = {
+  mode: ViewerMode;
+  message: string;
+  warning: string | null;
+  canResetCamera: boolean;
+};
+
+// ── Display & surface ────────────────────────────────────────────
+
+export type DisplayMode =
+  | "textured"
+  | "untextured"
+  | "wireframe"
+  | "texturedWireframe";
+
+export type ViewerSurfaceMode = "asset" | "texture";
+
+export type TextureViewMode = "rgb" | "rgba" | "r" | "g" | "b" | "alpha";
+
+// ── Scene context (Three.js handles) ─────────────────────────────
+
+export type SceneContext = {
+  renderer: WebGLRenderer;
+  scene: Scene;
+  camera: PerspectiveCamera;
+  controls: OrbitControls;
+  pmremGenerator: PMREMGenerator;
+  mountedObject: Group | Mesh | null;
+  sourceObject: Group | Mesh | null;
+  previewObject: Group | Mesh | null;
+  cleanupUrls: string[];
+  cleanupCallbacks: Array<() => void>;
+  mixer: AnimationMixer | null;
+  clips: AnimationClip[];
+  activeAction: AnimationAction | null;
+  mmdModel: MmdRuntimeModelHandle | null;
+  mmdMotion: MmdMotionPlayback | null;
+  textureRegistry: Map<string, Texture>;
+  rawMaxDimension: number;
+};
+
+// ── MMD runtime types ────────────────────────────────────────────
+
+export type MmdAnimationHandle = {
+  metadata: {
+    maxFrame?: number;
+  };
+};
+
+export type MmdRuntimeModelHandle = {
+  mesh: Object3D;
+  runtime?: {
+    reset(time: number): void;
+    setAnimation(animation: MmdAnimationHandle, mesh: Object3D): void;
+    tick(
+      time: number,
+      options: {
+        mesh: Object3D;
+        ik?: boolean;
+        physics?: boolean;
+      },
+    ): void;
+  };
+};
+
+export type MmdMotionPlayback = {
+  animation: MmdAnimationHandle;
+  duration: number;
+  currentTime: number;
+  label: string;
+};
+
+// ── Loaded preview ───────────────────────────────────────────────
+
+export type LoadedPreview = {
+  object: Group | Mesh;
+  cleanupUrls: string[];
+  cleanupCallbacks?: Array<() => void>;
+  clips: AnimationClip[];
+  formatVersion: string | null;
+  warnings?: string[];
+  lighting?: PreviewLightingPreset;
+  rendering?: PreviewRenderingPreset;
+  skipScaleNormalization?: boolean;
+  mmdMetadata?: MmdAssetMetadata;
+  mmdModel?: MmdRuntimeModelHandle;
+};
+
+export type LoadedMmdMotion = {
+  animation: MmdAnimationHandle;
+  duration: number;
+  label: string;
+};
+
+// ── Loading stage ────────────────────────────────────────────────
+
+export type DeferredTextureSnapshot = {
+  total: number;
+  loaded: number;
+  failed: number;
+  pending: number;
+  activeLabel: string | null;
+  bytes?: number;
+  readMs?: number;
+  parseMs?: number;
+};
+
+export type LoadingStageId =
+  | "scan"
+  | "resolve"
+  | "decode"
+  | "gpu"
+  | "scene"
+  | "ui";
+
+export type LoadingStageReporter = (stage: LoadingStageId) => void;
+
+export type LoadingStageSnapshot = {
+  activeStage: LoadingStageId;
+  activeStageStartedAt: number;
+  elapsedByStage: Partial<Record<LoadingStageId, number>>;
+  totalElapsedMs: number;
+};
+
+// ── Texture bundle ───────────────────────────────────────────────
+
+export type TextureBundle = {
+  albedo: Texture | null;
+  normal: Texture | null;
+  metalness: Texture | null;
+  roughness: Texture | null;
+  cleanupUrls: string[];
+};
+
+export type TextureSlotKey =
+  | "map"
+  | "normalMap"
+  | "metalnessMap"
+  | "roughnessMap"
+  | "emissiveMap"
+  | "alphaMap";
+
+export type TexturedMaterial = import("three").Material &
+  Partial<Record<TextureSlotKey, Texture | null>>;
+
+// ── Missing reference ────────────────────────────────────────────
+
+export type MissingReferenceError = Error & {
+  formatVersion: string | null;
+  missingPaths: string[];
+  unresolvedImages: string[];
+};
+
+// ── Preview support ──────────────────────────────────────────────
+
+export type PreviewSupportState =
+  | "implemented"
+  | "missingOptionalLoader"
+  | "unsupported";
+
+// ── Scene config (grid, scale, camera, texture filter) ───────────
+
+export type CameraPreset =
+  | "front"
+  | "back"
+  | "left"
+  | "right"
+  | "top"
+  | "bottom";
+
+export type GridConfig = {
+  cellSize: number;
+  label: string;
+  size: number;
+  divisions: number;
+};
+
+export type ScaleNormalizationResult = {
+  applied: boolean;
+  factor: number;
+  originalMaxDimension: number;
+  normalizedMaxDimension: number;
+  originalScale: import("three").Vector3 | null;
+};
+
+export type TextureFilterMode = "nearest" | "linear" | "trilinear";
+
+// ── Rendering & lighting presets ─────────────────────────────────
+
+export type PreviewRenderingPreset = {
+  logarithmicDepthBuffer: boolean;
+  outputColorSpace: ColorSpace;
+  toneMapping: ToneMapping;
+  toneMappingExposure: number;
+};
+
+export type PreviewLightingPreset = {
+  ambientIntensity: number;
+  keyIntensity: number;
+  keyPosition: readonly [number, number, number];
+  fillIntensity: number;
+};
+
+// ── Screenshot ───────────────────────────────────────────────────
+
+export type CanvasScreenshotOptions = {
+  mimeType?: "image/png" | "image/jpeg" | "image/webp";
+  quality?: number;
+  beforeCapture?: () => void;
+};
+
+export type CanvasScreenshot = {
+  dataUrl: string;
+  mimeType: string;
+  width: number;
+  height: number;
+};
+
+export type AssetViewportApi = {
+  captureScreenshot: (
+    options?: CanvasScreenshotOptions,
+  ) => Promise<CanvasScreenshot>;
+};
+
+// ── Loader registry ──────────────────────────────────────────────
+
+export type LoaderContext = {
+  renderer?: WebGLRenderer;
+  usdLoadPolicy?: import("./ipc").StageLoadPolicy;
+  variantSelections?: import("./ipc").VariantSelection[];
+  glbOverride?: ArrayBuffer | null;
+  onStage?: LoadingStageReporter;
+  onDeferredTexture?: (snapshot: DeferredTextureSnapshot) => void;
+  onWarning?: (warning: string) => void;
+};
+
+export type LoaderPlugin = {
+  id: string;
+  name: string;
+  extensions: readonly string[];
+  optional?: boolean;
+  installed?: boolean;
+  loadPreviewObject: (
+    file: import("./file").SelectedFile,
+    context: LoaderContext,
+  ) => Promise<LoadedPreview>;
+};
+
+export type RegisteredLoaderInfo = {
+  id: string;
+  name: string;
+  extension: string;
+  optional: boolean;
+  installed: boolean;
+};
+
+// ── Metadata collection ──────────────────────────────────────────
+
+export type MetadataCollection = {
+  metadata: AssetMetadata;
+  textureRegistry: Map<string, Texture>;
+};
+
+// ── Hierarchy ────────────────────────────────────────────────────
+
+export type HierarchyNode = {
+  name: string;
+  displayName?: string;
+  kind: string;
+  children: HierarchyNode[];
+  primPath?: string;
+};
+
+// ── Textures ─────────────────────────────────────────────────────
+
+export type TextureEntry = {
+  id: string;
+  label: string;
+  channel: string;
+  dimensions: string;
+  thumbnailUrl: string | null;
+  previewFlipY?: boolean;
+  sourceKind: "embedded" | "external" | "standalone" | "unresolved" | "unknown";
+};
+
+// ── Material texture slot ────────────────────────────────────────
+
+export type MaterialTextureSlot = {
+  name: string;
+};
+
+// ── MMD material ─────────────────────────────────────────────────
+
+export type MmdMaterialEntry = {
+  materialIndex: number | null;
+  name: string;
+  englishName: string | null;
+  diffuse: [number, number, number, number] | null;
+  specular: [number, number, number] | null;
+  ambient: [number, number, number] | null;
+  specularPower: number | null;
+  edgeColor: [number, number, number, number] | null;
+  edgeSize: number | null;
+  texturePath: string | null;
+  sphereTexturePath: string | null;
+  sphereMode: string | null;
+  toonTexturePath: string | null;
+  sharedToonIndex: number | null;
+  transparencyMode: string | null;
+  renderOrderBucket: string | null;
+  faceCount: number | null;
+  flags: Record<string, boolean> | null;
+  unsupportedDrawFlags: string[];
+};
+
+// ── Material entry ───────────────────────────────────────────────
+
+export type MaterialEntry = {
+  id: string;
+  name: string;
+  type: string;
+  color: string | null;
+  opacity: number;
+  transparent: boolean;
+  textureCount: number;
+  boundMeshes: string[];
+  baseColorFactor: [number, number, number, number] | null;
+  metallicFactor: number | null;
+  roughnessFactor: number | null;
+  emissiveFactor: [number, number, number] | null;
+  baseColorTexture: MaterialTextureSlot | null;
+  metallicRoughnessTexture: MaterialTextureSlot | null;
+  normalTexture: MaterialTextureSlot | null;
+  emissiveTexture: MaterialTextureSlot | null;
+  alphaMode: "OPAQUE" | "MASK" | "BLEND" | "unknown";
+  usdPrimPath: string | null;
+  mmd: MmdMaterialEntry | null;
+};
+
+// ── Light entry ──────────────────────────────────────────────────
+
+export type LightEntry = {
+  id: string;
+  name: string;
+  type: string;
+  color: string | null;
+  intensity: number;
+};
+
+// ── Camera entry ─────────────────────────────────────────────────
+
+export type CameraEntry = {
+  id: string;
+  name: string;
+  projection: "perspective" | "orthographic";
+  fov: number | null;
+  aspect: number | null;
+  near: number;
+  far: number;
+};
+
+// ── MMD asset metadata ───────────────────────────────────────────
+
+export type MmdSectionEntry = {
+  name: string;
+  count: number;
+  offset: number;
+  byteLength: number;
+};
+
+export type MmdAssetMetadata = {
+  format: "pmx" | "pmd";
+  version: number;
+  encoding: string | null;
+  name: string;
+  englishName: string;
+  comment: string;
+  englishComment: string;
+  counts: Record<string, number>;
+  additionalUvCount: number | null;
+  indexSizes: Record<string, number> | null;
+  trailingBytes: number;
+  sections: MmdSectionEntry[];
+  diagnostics: Array<{
+    level: "warning" | "error";
+    code: string;
+    message: string;
+  }>;
+};
+
+export type MmdBoneEntry = {
+  boneIndex: number | null;
+  parentIndex: number | null;
+  parentName: string | null;
+  name: string | null;
+  englishName: string | null;
+  restPosition: [number, number, number] | null;
+  layer: number | null;
+  appendTransform: {
+    parentIndex: number;
+    parentName: string | null;
+    weight: number;
+  } | null;
+  flags: Record<string, boolean> | null;
+  ik: {
+    roles: string[];
+    goalBoneIndex: number | null;
+    effectorBoneIndex: number | null;
+    iterationCount: number | null;
+    maxAnglePerIteration: number | null;
+    linkCount: number | null;
+    limitKinds: string[];
+  } | null;
+};
+
+export type MmdMorphEntry = {
+  name: string | null;
+  englishName: string | null;
+  type: string | null;
+  boneOffsetCount: number;
+  groupOffsetCount: number;
+  flipOffsetCount: number;
+  impulseOffsetCount: number;
+};
+
+// ── Object info ──────────────────────────────────────────────────
+
+export type ObjectInfo = {
+  name: string;
+  kind: string;
+  visible: boolean;
+  position: [number, number, number];
+  rotation: [number, number, number];
+  scale: [number, number, number];
+  boundingBox: [number, number, number, number, number, number] | null;
+  vertexCount: number | null;
+  triangleCount: number | null;
+  materialNames: string[];
+  materialIds: string[];
+  morphTargets: MorphTargetEntry[];
+  childCount: number | null;
+  animatesWithClips: string[];
+  userData: Record<string, unknown> | null;
+  mmdBone: MmdBoneEntry | null;
+};
+
+export type MorphTargetEntry = {
+  index: number;
+  name: string;
+  value: number;
+  mmd: MmdMorphEntry | null;
+};
+
+// ── Top-level asset metadata ─────────────────────────────────────
+
+export type AssetMetadata = {
+  formatLabel: string;
+  formatVersion: string | null;
+  nodeCount: number;
+  meshCount: number;
+  materialCount: number;
+  textureCount: number;
+  hasAnimation: boolean;
+  hierarchy: HierarchyNode[];
+  textures: TextureEntry[];
+  materials: MaterialEntry[];
+  lights: LightEntry[];
+  cameras: CameraEntry[];
+  objectInfo: Record<string, ObjectInfo>;
+  mmd?: MmdAssetMetadata;
+};
+
+// ── AssetViewport viewer settings ────────────────────────────────
+
+export type BackgroundPreset = "gray" | "charcoal" | "light";
+
+export type CameraPresetRequest = {
+  preset: CameraPreset;
+  version: number;
+};
+
+export type EnvironmentPreset = "studio" | "neutral" | "outdoor";
+
+export type ToneMappingMode = "linear" | "aces" | "reinhard";
+
+// ── Image toolbar ────────────────────────────────────────────────
+
+export type TextureColorSpace = "srgb" | "linear" | "raw";
+
+export type BuildImageToolbarOptions = {
+  channelMode: string | null;
+  channelOptions: Array<{ id: string; label: string }>;
+  onSelectChannel?: (mode: string) => void;
+  colorSpace: TextureColorSpace;
+  onSelectColorSpace?: (mode: TextureColorSpace) => void;
+  exposure: number;
+  bgMode: string;
+  onSelectBgMode?: (mode: string) => void;
+  tilingMode: string;
+  onSelectTilingMode?: (mode: string) => void;
+  tileCount: number;
+};
+
+// ── 3D toolbar ───────────────────────────────────────────────────
+
+export type Build3DToolbarOptions = {
+  cameraPreset: string | null;
+  cameraPresetOptions: Array<{ id: string; label: string }>;
+  onSelectCameraPreset?: (preset: string) => void;
+  onCycleCamera?: () => void;
+  showTexture: boolean;
+  onToggleTexture: () => void;
+  showUnlit: boolean;
+  onToggleUnlit: () => void;
+  showNormals?: boolean;
+  onToggleNormals?: () => void;
+  showVertexColors?: boolean;
+  onToggleVertexColors?: () => void;
+  showWireframe: boolean;
+  onToggleWireframe: () => void;
+  environmentPreset: string;
+  environmentPresetOptions: Array<{ id: string; label: string }>;
+  onSelectEnvironmentPreset?: (preset: string) => void;
+  showShadows?: boolean;
+  onToggleShadows?: () => void;
+  showEnvironmentBackground?: boolean;
+  onToggleEnvironmentBackground?: () => void;
+  showBoundingBoxes?: boolean;
+  onToggleBoundingBoxes?: () => void;
+  showSkeleton?: boolean;
+  onToggleSkeleton?: () => void;
+};
+
+// ── Animation state ──────────────────────────────────────────────
+
+export type AnimationState = {
+  clipNames: string[];
+  activeClipIndex: number;
+  currentTime: number;
+  duration: number;
+  isPlaying: boolean;
+};
